+++ examples/data-sources/onepasswordorg_group/data-source.tf
@@ -1,0 +1,7 @@
+data "onepasswordorg_group" "test" {
+  name = "test-group"
+}
+
+output "group_test" {
+  value = data.onepasswordorg_group.test
+}