+++ examples/data-sources/onepasswordorg_user/data-source.tf
@@ -2,6 +2,6 @@ data "onepasswordorg_user" "test" {
   email = "user0@slok.dev"
 }
 
-output "user_test_id" {
+output "user_test" {
   value = data.onepasswordorg_user.test
 }